% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scilicium.R
\name{display_scilicium_all}
\alias{display_scilicium_all}
\title{Display all scilicium palettes}
\usage{
display_scilicium_all()
}
\value{
A custom ggplot2 color scale
}
\description{
Display all Scilicium palettes
}
