% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scilicium.R
\name{generate_seq_palette}
\alias{generate_seq_palette}
\title{Helper function that helps generate a sequential palette of arbitrary length}
\usage{
generate_seq_palette(name = "LightGreen", n = 5)
}
\arguments{
\item{name}{The name of the palette to use}

\item{n}{The number of colors in the sequential palette}
}
\value{
A vector of RGB codes
}
\description{
Helper function that helps generate a sequential palette of arbitrary length
}
