% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scilicium.R
\name{scale_fill_scilicium}
\alias{scale_fill_scilicium}
\title{SciLicium scale fill for ggplot2}
\usage{
scale_fill_scilicium(name = "SciliciumLight")
}
\arguments{
\item{name}{The name of the palette to use. Currently available palettes are "SciliciumLight" (default, 5 colors), "SciliciumDark" (5 colors), "SciliciumPaired" (8 colors), "KombuGreen" (5 colors) and "EarthYellow" (5 colors)}
}
\value{
A custom ggplot2 color scale
}
\description{
SciLicium scale fill
}
\examples{
scale_fill_scilicium("KombuGreen")
}
