% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scilicium.R
\name{scale_fill_scilicium_gradients}
\alias{scale_fill_scilicium_gradients}
\title{SciLicium scale gradient fill for ggplot2}
\usage{
scale_fill_scilicium_gradients(
  n = 5,
  color = "green",
  shade = c("light", "dark")
)
}
\arguments{
\item{n}{An integer representing the number of colors in the gradient}

\item{color}{A color to use (currently only "green" is supported)}

\item{shade}{A shade to use ("light" or "dark")}
}
\value{
A custom ggplot2 color scale
}
\description{
SciLicium gradient colors
}
\examples{
scale_fill_scilicium_gradients(10, "green", "dark")
}
