% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scilicium.R
\name{scilicium_pal}
\alias{scilicium_pal}
\title{Return a scilicium palette}
\usage{
scilicium_pal(n, name)
}
\arguments{
\item{n}{An integer representing the number of colors in the palette}

\item{name}{The name of the palette to use. Currently available palettes are SciliciumLight, SciliciumDark, SciliciumPaired2, SciliciumPairedLight, SciliciumPairedDark, KombuGreen and EarthYellow (for qualitative palettes), DarkGreen, LightGreen, DarkBlue, LightBlue, DarkRed, LightRed, DarkOcre, LightOcre (for sequential palettes)}
}
\value{
A vector of html color codes
}
\description{
Access a Scilicium palette
}
\examples{
scilicium_pal(5, "DarkGreen")
scilicium_pal(10, "SciliciumPairedLight")
}
