% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scilicium.R
\name{theme_genolens}
\alias{theme_genolens}
\title{Scilicium/genolens theme for ggplot2}
\usage{
theme_genolens()
}
\description{
Function to use the ggplot2 SciLicium genolens theme.
}
\examples{
library(ggplot2)
data("midwest", package = "ggplot2")
# midwest <- read.csv("http://goo.gl/G1K41K")  # bkup data source

# Scatterplot
gg <- ggplot(midwest, aes(x=area, y=poptotal)) +
  geom_point(aes(col=state, size=popdensity)) +
  geom_smooth(method="loess", se=FALSE) +
  xlim(c(0, 0.1)) +
  ylim(c(0, 500000)) +
  labs(subtitle="Area Vs Population",
       y="Population",
       x="Area",
       title="Scatterplot",
       caption = "Source: midwest")

gg

ggsci <- gg + scale_color_scilicium("GenolensLong1") + theme_genolens()
ggsci
}
