% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scilicium.R
\name{display_scilicium_all}
\alias{display_scilicium_all}
\title{Display all scilicium palettes}
\usage{
display_scilicium_all()
}
\value{
A very nice plot!
}
\description{
Display all available Scilicium palettes
}
\seealso{
Other palette vizualization functions: 
\code{\link{display_scilicium_pal}()}
}
\concept{palette vizualization functions}
