% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scilicium.R
\name{scale_fill_scilicium}
\alias{scale_fill_scilicium}
\title{SciLicium scale fill for ggplot2}
\usage{
scale_fill_scilicium(n = 5, name = "SciliciumLight")
}
\arguments{
\item{n}{An integer representing the number of colors in the palette}

\item{name}{The name of the palette to use. Currently available palettes are SciliciumLight, SciliciumDark, SciliciumPaired2, SciliciumPairedLight, SciliciumPairedDark, KombuGreen and EarthYellow (for qualitative palettes), DarkGreen, LightGreen, DarkBlue, LightBlue, DarkRed, LightRed, DarkOcre, LightOcre (for sequential palettes)}
}
\value{
A custom ggplot2 color fill scale
}
\description{
SciLicium scale fill
}
\examples{
scale_fill_scilicium(5, "KombuGreen")
library(ggplot2)
n_category <- 5
n_replicates <- 5
tmp_df <- data.frame(x = runif(n_replicates * n_category),
  y=runif(n_replicates * n_category),
  category=letters[rep(1:n_category, n_replicates)])
gg <- ggplot(tmp_df, aes(x=category, y=y, fill=category)) +
  geom_boxplot() +
  scale_fill_scilicium(5, "DarkRed") +
  theme_scilicium()
}
