% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scilicium.R
\name{scale_fill_scilicium}
\alias{scale_fill_scilicium}
\title{SciLicium scale fill for ggplot2}
\usage{
scale_fill_scilicium(theme = "SciliciumLight", pal_key = scilicium_palettes)
}
\arguments{
\item{theme}{A theme to use}

\item{pal_key}{A list of palettes}
}
\value{
A custom ggplot2 color scale
}
\description{
SciLicium scale fill
}
\examples{
scale_fill_scilicium(theme="KombuGreen")
}
