% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scilicium.R
\name{scale_fill_scilicium_gradients}
\alias{scale_fill_scilicium_gradients}
\title{SciLicium scale gradient fill for ggplot2}
\usage{
scale_fill_scilicium_gradients(
  n = 5,
  name = c("DarkGreen", "DarkBlue", "DarkRed", "DarkOCre", "LightGreen", "LightBlue",
    "LightRed", "LightOcre")
)
}
\arguments{
\item{n}{An integer representing the number of colors in the gradient}

\item{name}{The name of a sequential palette to use (currently only "DarkGreen", "DarkBlue", "DarkRed", "DarkOCre", "LightGreen", "LightBlue", "LightRed", "LightOcre" are supported)}
}
\value{
A custom ggplot2 color scale
}
\description{
SciLicium gradient colors
}
\examples{
scale_fill_scilicium_gradients(10, "DarkGreen")
library(ggplot2)
n_category <- 5
n_replicates <- 5
tmp_df <- data.frame(x = runif(n_replicates * n_category),
  y=runif(n_replicates * n_category),
  category=letters[rep(1:n_category, n_replicates)])
gg <- ggplot(tmp_df, aes(x=category, y=y, fill=category)) +
  geom_boxplot() +
  scale_fill_scilicium_gradients(5, "DarkRed") +
  theme_scilicium()

}
